Removed the QSort function from the main file to the module and added the header
