+++ modulesAndFiles/modulesAndFiles/sorting.h
@@ -1,1 +1,3 @@
 #pragma once
+
+void qSort(int* originalArray, int size, int beginning, int ending); // sorting the array using QSort
